tests/wpt/tests/css/css-overflow/parsing/max-lines-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overflow: parsing max-lines with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow/#propdef-max-lines">
<meta name="assert" content="max-lines supports the full grammar 'none | <integer>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("max-lines", 'none');
test_valid_value("max-lines", '1');
test_valid_value("max-lines", '6');
</script>
</body>
</html>
